streamlit/requirements.txt
streamlit>=1.44.0
plotly>=6.0.0
pandas>=2.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
